cleaning out and moving server side files
