--- client/index.js
@@ -1,3 +1,0 @@
-/*
-GET TO THE CHOPPA
-*/
