--- client/login.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title>YumSnap!</title>
-    <!-- change this up! http://www.bootstrapcdn.com/bootswatch/ -->
-    <link href="https://maxcdn.bootstrapcdn.com/bootswatch/3.3.6/cosmo/bootstrap.min.css" type="text/css" rel="stylesheet"/>
-  </head>
-
-  <body>
-    <div id="app"></div>
-    <div class="container">
-
-    <div class="jumbotron text-center">
-        <h1><span class="fa fa-lock"></span> Yum Snap</h1>
-
-        <p>Login or Register with:</p>          
-
-        <a href="/login" class="btn btn-default"><span class="fa fa-user"></span> Local Login</a>
-        <a href="/signup" class="btn btn-default"><span class="fa fa-user"></span> Local Signup</a>
-        <a href="/auth/facebook" class="btn btn-primary"><span class="fa fa-facebook"></span> Facebook</a>
-        <a href="/auth/twitter" class="btn btn-info"><span class="fa fa-twitter"></span> Twitter</a>
-
-
-
-    </div>
-
-</div>
-    <script src="client.min.js"></script>
-  </body>
-</html>
