+++ server/database/db.js
@@ -1,4 +1,4 @@
-var config      = require('../knexfile.js');  
+var config      = require('./knexfile');  
 var env         = 'development';  
 var knex        = require('knex')(config[env]);
